Simplified layout & service
