--- src/MudBlazor.Docs/Pages/Getting Started/_Imports.razor
@@ -1,1 +1,0 @@
-﻿@layout DocsStartedLayout
