--- src/MudBlazor.Docs/Pages/Mud/_Imports.razor
@@ -1,1 +1,0 @@
-﻿@layout DocsMoreLayout
