+++ src/MudBlazor.Docs/Shared/Appbar.razor
@@ -1,5 +1,5 @@
 ﻿<MudHidden Breakpoint="Breakpoint.MdAndUp">
-    <MudIconButton OnClick="ToggleDrawerState" Icon="@Icons.Rounded.Notes" Color="Color.Inherit" Edge="Edge.Start"/>
+    <MudIconButton OnClick="DrawerToggleCallback" Icon="@Icons.Rounded.Notes" Color="Color.Inherit" Edge="Edge.Start"/>
     <MudSpacer/>
     <NavLink ActiveClass="d-flex" href="/">
         <MudBlazorLogo Class="docs-mudblazor-logo"/>
