+++ src/MudBlazor.Docs/Shared/DocsLayout.razor.cs
@@ -4,16 +4,44 @@
 
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
+using MudBlazor.Docs.Services;
 
 namespace MudBlazor.Docs.Shared;
 
 public partial class DocsLayout : LayoutComponentBase
 {
+    [Inject] private LayoutService LayoutService { get; set; }
+    [Inject] private NavigationManager NavigationManager { get; set; }
+    
     private NavMenu _navMenuRef;
-
+    private bool _drawerOpen = true;
+    private bool _topMenuOpen = false;
+    protected override void OnInitialized()
+    {
+        LayoutService.SetBaseTheme(Theme.DocsTheme());
+    }
+    
     protected override void OnAfterRender(bool firstRender)
     {
         //refresh nav menu because no parameters change in nav menu but internal data does
         _navMenuRef?.Refresh();
     }
+
+    private void ToggleDrawer()
+    {
+        _drawerOpen = !_drawerOpen;
+    }
+
+    private void OpenTopMenu()
+    {
+        _topMenuOpen = true;
+    }
+
+    private void OnDrawerOpenChanged(bool value)
+    {
+        _topMenuOpen = false;
+        _drawerOpen = value;
+        StateHasChanged();
+    }
+
 }
