--- src/MudBlazor.Docs/Shared/DocsMoreLayout.razor
@@ -1,21 +1,0 @@
-﻿@inherits LayoutComponentBase
-@layout DocsMainLayout
-
-<Drawer>
-    <MudText Typo="Typo.subtitle1" Class="mt-1 mb-2 ml-2 d-flex d-md-none"><b>Discover More</b></MudText>
-    <MudNavLink Href="/mud/introduction">What is MudBlazor?</MudNavLink>
-    <MudNavGroup Title="Community" Expanded="true" HideExpandIcon="true">
-        <MudNavLink Href="/mud/community/getting-help">Getting Help</MudNavLink>
-        <MudNavLink Href="/mud/community/reporting-bugs">Reporting Bugs</MudNavLink>
-        <MudNavLink Href="/mud/community/contribution">Contribution</MudNavLink>
-    </MudNavGroup>
-    <MudNavGroup Title="Project" Expanded="true" HideExpandIcon="true">
-        <MudNavLink Href="/mud/announcements">Announcements</MudNavLink>
-        <MudNavLink Href="/mud/project/how-it-started">How it started</MudNavLink>
-        <MudNavLink Href="/mud/project/releases">Releases</MudNavLink>
-        <MudNavLink Href="/mud/project/roadmap">Roadmap</MudNavLink>
-        <MudNavLink Href="/mud/project/sponsor">Sponsors & Backers</MudNavLink>
-        <MudNavLink Href="/mud/project/team">Team & Contributors</MudNavLink>
-    </MudNavGroup>
-</Drawer>
-@Body
