--- src/MudBlazor.Docs/Shared/DocsStartedLayout.razor
@@ -1,11 +1,0 @@
-﻿@inherits LayoutComponentBase
-@layout DocsMainLayout
-
-<Drawer>
-    <MudText Typo="Typo.subtitle1" Class="mt-1 mb-2 ml-2 d-flex d-md-none"><b>Getting Started</b></MudText>
-    <MudNavLink Href="/getting-started/installation">Installation</MudNavLink>
-    <MudNavLink Href="/getting-started/layouts">Layouts</MudNavLink>
-    <MudNavLink Href="/getting-started/usage">Usage</MudNavLink>
-    <MudNavLink Href="/getting-started/wireframes">Wireframes</MudNavLink>
-</Drawer>
-@Body
