--- src/MudBlazor.Docs/Shared/Drawer.razor
@@ -1,32 +1,0 @@
-﻿<MudDrawer Open="@LayoutService.DrawerIsOpen" OpenChanged="LayoutService.SetDrawer" ClipMode="DrawerClipMode.Docked" Elevation="0" Breakpoint="Breakpoint.Lg">
-    <MudHidden Breakpoint="Breakpoint.MdAndUp">
-        <MudToolBar Dense="true" DisableGutters="true" Class="docs-gray-bg">
-            @if (_topMenuOpen == false)
-            {
-                <MudIconButton Icon="@Icons.Rounded.ArrowBack" Color="Color.Inherit" OnClick="@((e) => OpenTopMenu())"/>
-            }
-            else
-            {
-                <MudIconButton Icon="@Icons.Rounded.Close" Color="Color.Inherit" OnClick="LayoutService.ToggleDrawer"/>
-            }
-            <MudSpacer/>
-            <AppbarButtons/>
-        </MudToolBar>
-
-        @if (_topMenuOpen == true)
-        {
-            <TopNavMenu />
-        }
-        else
-        {
-            <MudNavMenu Color="Color.Primary" Rounded="true" Dense="true" Margin="Margin.Dense" Class="pa-2 overflow-auto mb-3">
-                @ChildContent
-            </MudNavMenu>
-        }
-    </MudHidden>
-    <MudHidden Breakpoint="Breakpoint.MdAndUp" Invert="true">
-        <MudNavMenu Color="Color.Primary" Rounded="true" Dense="true" Margin="Margin.Dense" Class="pa-2 overflow-auto mb-3">
-            @ChildContent
-        </MudNavMenu>
-    </MudHidden>
-</MudDrawer>
