+++ src/MudBlazor.Docs/Shared/LandingLayout.razor
@@ -3,15 +3,29 @@
 
 <MudLayout Class="mudblazor-landingpage">
     <MudAppBar Class="landing-appbar" Elevation="0">
-        <Appbar/>
+        <Appbar DrawerToggleCallback="ToggleDrawer" />
     </MudAppBar>
-    <MudDrawer Open="@LayoutService.DrawerIsOpen" OpenChanged="@LayoutService.SetDrawer" Elevation="25" Variant="@DrawerVariant.Temporary">
+    <MudDrawer @bind-Open="@_drawerOpen" Elevation="25" Variant="@DrawerVariant.Temporary">
         <MudToolBar Dense="true" DisableGutters="true" Class="px-1 docs-gray-bg">
-            <MudIconButton Icon="@Icons.Rounded.Close" Color="Color.Inherit" OnClick="@LayoutService.ToggleDrawer"/>
+            <MudIconButton Icon="@Icons.Rounded.Close" Color="Color.Inherit" OnClick="@ToggleDrawer"/>
             <MudSpacer/>
             <AppbarButtons/>
         </MudToolBar>
-        <TopNavMenu/>
+        <MudNavMenu Color="Color.Primary" Margin="Margin.Dense" Rounded="true" Class="pa-2">
+            <MudNavLink Href="/components/alert">Docs</MudNavLink>
+            <MudNavLink Href="/getting-started/installation">Getting Started</MudNavLink>
+            <MudNavLink Href="/mud/introduction" Match="NavLinkMatch.All">Discover More</MudNavLink>
+            <MudNavGroup Title="Products" ExpandIcon="@Icons.Material.Filled.ExpandMore">
+                 <MudNavLink Href="https://mudblazor.com/">MudBlazor</MudNavLink>
+                 <MudNavLink Href="https://try.mudblazor.com/">TryMudBlazor</MudNavLink>
+                 <MudNavLink Href="https://github.com/MudBlazor/ThemeManager">ThemeManager</MudNavLink>
+                 <MudNavLink Href="https://github.com/MudBlazor/Templates">Templates</MudNavLink>
+                <MudListSubheader Class="mt-4 pl-8 pb-2"><b>Coming soon</b></MudListSubheader>
+                 <MudNavLink Disabled="true">MudStore</MudNavLink>
+                 <MudNavLink Disabled="true">MudAcademy</MudNavLink>
+                 <MudNavLink Disabled="true">MudExtensions</MudNavLink>
+            </MudNavGroup>
+        </MudNavMenu>
     </MudDrawer>
     <MudMainContent Class="pa-0">
         @Body
