+++ src/MudBlazor.Docs/Shared/NavMenu.razor
@@ -1,6 +1,7 @@
 @using System.Linq
 @using MudBlazor.Docs.Extensions
 
+
 <MudText Typo="Typo.subtitle1" Class="mt-1 mb-2 ml-2 d-flex d-md-none"><b>Docs</b></MudText>
 <MudNavGroup Title="Components" Icon="@Icons.Material.Filled.Dashboard" Expanded="@(_section == "components")" ExpandIcon="@Icons.Material.Filled.ExpandMore">
     @foreach (var item in MenuService.Components)
@@ -34,10 +35,10 @@
         <MudNavLink Href="@link.Href">@link.Title</MudNavLink>
     }
 </MudNavGroup>
-<MudNavGroup Title="Customization" Icon="@Icons.Material.Filled.Handyman" Expanded="@(_section == "customization")" ExpandIcon="@Icons.Material.Filled.ExpandMore">
-    <MudNavLink Href="customization/default-theme">Default Theme</MudNavLink>
-    @foreach (var link in MenuService.Customization)
-    {
-        <MudNavLink Href="@link.Href">@link.Title</MudNavLink>
-    }
-</MudNavGroup>
+ <MudNavGroup Title="Customization" Icon="@Icons.Material.Filled.Handyman" Expanded="@(_section == "customization")" ExpandIcon="@Icons.Material.Filled.ExpandMore">
+     <MudNavLink Href="customization/default-theme">Default Theme</MudNavLink>
+     @foreach (var link in MenuService.Customization)
+     {
+         <MudNavLink Href="@link.Href">@link.Title</MudNavLink>
+     }
+ </MudNavGroup>
