--- src/MudBlazor.Docs/Shared/TopNavMenu.razor
@@ -1,15 +1,0 @@
-﻿<MudNavMenu Color="Color.Primary" Margin="Margin.Dense" Rounded="true" Class="pa-2">
-    <MudNavLink Href="/components/alert">Docs</MudNavLink>
-    <MudNavLink Href="/getting-started/installation">Getting Started</MudNavLink>
-    <MudNavLink Href="/mud/introduction" Match="NavLinkMatch.All">Discover More</MudNavLink>
-    <MudNavGroup Title="Products" ExpandIcon="@Icons.Material.Filled.ExpandMore">
-         <MudNavLink Href="https://mudblazor.com/">MudBlazor</MudNavLink>
-         <MudNavLink Href="https://try.mudblazor.com/">TryMudBlazor</MudNavLink>
-         <MudNavLink Href="https://github.com/MudBlazor/ThemeManager">ThemeManager</MudNavLink>
-         <MudNavLink Href="https://github.com/MudBlazor/Templates">Templates</MudNavLink>
-        <MudListSubheader Class="mt-4 pl-8 pb-2"><b>Coming soon</b></MudListSubheader>
-         <MudNavLink Disabled="true">MudStore</MudNavLink>
-         <MudNavLink Disabled="true">MudAcademy</MudNavLink>
-         <MudNavLink Disabled="true">MudExtensions</MudNavLink>
-    </MudNavGroup>
-</MudNavMenu>
