+++ src/MudBlazor.Docs/Shared/Utilities/WireframeLayout.razor
@@ -1,4 +1,5 @@
-﻿@inherits LayoutComponentBase
+﻿@namespace MudBlazor.Docs.Shared
+@inherits LayoutComponentBase
 
 @Body
